Move auth keeper benchmarks to sep. file
